+++ api/openapi_client/README.md
@@ -5,7 +5,7 @@ A client library for accessing HuBMAP Ontology API
 First, create a client:
 
 ```python
-from hu_bmap_ontology_api_client import Client
+from openapi_client import Client
 
 client = Client(base_url="https://api.example.com")
 ```
@@ -13,7 +13,7 @@ client = Client(base_url="https://api.example.com")
 If the endpoints you're going to hit require authentication, use `AuthenticatedClient` instead:
 
 ```python
-from hu_bmap_ontology_api_client import AuthenticatedClient
+from openapi_client import AuthenticatedClient
 
 client = AuthenticatedClient(base_url="https://api.example.com", token="SuperSecretToken")
 ```
@@ -21,9 +21,9 @@ client = AuthenticatedClient(base_url="https://api.example.com", token="SuperSec
 Now call your endpoint and use your models:
 
 ```python
-from hu_bmap_ontology_api_client.models import MyDataModel
-from hu_bmap_ontology_api_client.api.my_tag import get_my_data_model
-from hu_bmap_ontology_api_client.types import Response
+from openapi_client.models import MyDataModel
+from openapi_client.api.my_tag import get_my_data_model
+from openapi_client.types import Response
 
 my_data: MyDataModel = get_my_data_model.sync(client=client)
 # or if you need more info (e.g. status_code)
@@ -33,9 +33,9 @@ response: Response[MyDataModel] = get_my_data_model.sync_detailed(client=client)
 Or do the same thing with an async version:
 
 ```python
-from hu_bmap_ontology_api_client.models import MyDataModel
-from hu_bmap_ontology_api_client.api.my_tag import get_my_data_model
-from hu_bmap_ontology_api_client.types import Response
+from openapi_client.models import MyDataModel
+from openapi_client.api.my_tag import get_my_data_model
+from openapi_client.types import Response
 
 my_data: MyDataModel = await get_my_data_model.asyncio(client=client)
 response: Response[MyDataModel] = await get_my_data_model.asyncio_detailed(client=client)
@@ -70,7 +70,7 @@ Things to know:
 
 1. All path/query params, and bodies become method arguments.
 1. If your endpoint had any tags on it, the first tag will be used as a module name for the function (my_tag above)
-1. Any endpoint which did not have a tag will be in `hu_bmap_ontology_api_client.api.default`
+1. Any endpoint which did not have a tag will be in `openapi_client.api.default`
 
 ## Building / publishing this Client
 This project uses [Poetry](https://python-poetry.org/) to manage dependencies  and packaging.  Here are the basics:
