+++ api/openapi_client/openapi_client/api/default/semantics_semantic_id_semantics_get.py
