+++ api/openapi_client/openapi_client/api/default/tui_tui_id_semantics_get.py
