+++ api/openapi_client/openapi_client/models/concept_detail.py
