+++ api/openapi_client/openapi_client/models/semantic_stn.py
