+++ api/openapi_client/openapi_client/py.typed
