+++ api/openapi_client/pyproject.toml
@@ -1,15 +1,15 @@
 [tool.poetry]
-name = "hu-bmap-ontology-api-client"
+name = "openapi_client"
 version = "1.1.1"
-description = "A client library for accessing HuBMAP Ontology API"
+description = "A client library for accessing UBKG API"
 
 authors = []
 
 readme = "README.md"
 packages = [
-    {include = "hu_bmap_ontology_api_client"},
+    {include = "openapi_client"},
 ]
-include = ["CHANGELOG.md", "hu_bmap_ontology_api_client/py.typed"]
+include = ["CHANGELOG.md", "openapi_client/py.typed"]
 
 [tool.poetry.dependencies]
 python = "^3.6"
